requirements.txt
python-dotenv
crewai>=0.35.0
crewai-tools>=0.2.0   
langchain>=0.2.14
langchain-community>=0.2.11
langchain-huggingface
huggingface-hub>=0.26.0
transformers>=4.44.0
sentence-transformers>=3.0.1
requests
beautifulsoup4
duckduckgo-search>=5.3.0
ddgs>=2.1.2
pandas
numpy
ipykernel
notebook
litellm